Update and debugged controller, service, and route for deleting user

+++ src/services/deleteUserService.js
@@ -2,16 +2,19 @@
  * Service layer responsbile for with all models to delete user data 
  */
 
-const services = require{"./index"}
+const services = require("./index")
 
 async function deleteAllData(userId){
     for (const serviceName in services){
         const service = services[serviceName];
+
+        if (serviceName === "deleteUserService") continue;
+
+        if(typeof service.deleteUser === "function"){
+            await service.deleteUser(userId);
+        }
     }
 
-    if(typeof service.deleteUser === "function"){
-        await service.deleteUser(userId);
-    }
 } 
 
 module.exports = { deleteAllData }
